ai_news_bot/requirements.txt
feedparser>=6.0.10
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0
schedule>=1.2.0
aiohttp>=3.8.5
aiodns>=3.0.0
apscheduler>=3.10.4
